Update FavExchangeList screen components to simplify tests
